--- model/entity/login_session.go
@@ -1,8 +1,0 @@
-package entity
-
-import "time"
-
-// LoginSession ログインに必要なセッション情報を保持
-type LoginSession struct {
-	LastLoginDate time.Time
-}
